backend/requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
pydantic
aiosqlite
